+++ Funcky/Monads/Option/Option.Core.cs
@@ -3,7 +3,7 @@ using System.Diagnostics.Contracts;
 
 namespace Funcky.Monads
 {
-    public readonly partial struct Option<TItem> : IToString
+    public readonly partial struct Option<TItem>
         where TItem : notnull
     {
         private readonly bool _hasItem;