+++ Funcky/Functional/ActionToUnit.cs
@@ -1,9 +1,11 @@
-﻿using System;
+using System;
+using System.Diagnostics.Contracts;
 
 namespace Funcky
 {
     public static partial class Functional
     {
+        [Pure]
         public static Func<Unit> ActionToUnit(Action action)
             => () =>
             {
@@ -11,6 +13,7 @@ namespace Funcky
                 return default;
             };
 
+        [Pure]
         public static Func<T, Unit> ActionToUnit<T>(Action<T> action)
         {
             return parameter =>
@@ -20,6 +23,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, Unit> ActionToUnit<T1, T2>(Action<T1, T2> action)
         {
             return (p1, p2) =>
@@ -29,6 +33,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, T3, Unit> ActionToUnit<T1, T2, T3>(Action<T1, T2, T3> action)
         {
             return (p1, p2, p3) =>
@@ -38,6 +43,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, T3, T4, Unit> ActionToUnit<T1, T2, T3, T4>(Action<T1, T2, T3, T4> action)
         {
             return (p1, p2, p3, p4) =>
@@ -47,6 +53,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, T3, T4, T5, Unit> ActionToUnit<T1, T2, T3, T4, T5>(Action<T1, T2, T3, T4, T5> action)
         {
             return (p1, p2, p3, p4, p5) =>
@@ -56,6 +63,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, T3, T4, T5, T6, Unit> ActionToUnit<T1, T2, T3, T4, T5, T6>(Action<T1, T2, T3, T4, T5, T6> action)
         {
             return (p1, p2, p3, p4, p5, p6) =>
@@ -65,6 +73,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, T3, T4, T5, T6, T7, Unit> ActionToUnit<T1, T2, T3, T4, T5, T6, T7>(Action<T1, T2, T3, T4, T5, T6, T7> action)
         {
             return (p1, p2, p3, p4, p5, p6, p7) =>
@@ -74,6 +83,7 @@ namespace Funcky
             };
         }
 
+        [Pure]
         public static Func<T1, T2, T3, T4, T5, T6, T7, T8, Unit> ActionToUnit<T1, T2, T3, T4, T5, T6, T7, T8>(Action<T1, T2, T3, T4, T5, T6, T7, T8> action)
         {
             return (p1, p2, p3, p4, p5, p6, p7, p8) =>