+++ changelog.md
@@ -35,3 +35,4 @@
 * Add `IndexOfOrNone`, `LastIndexOfOrNone`, `IndexOfAnyOrNone` and `LastIndexOfAnyOrNone` extension methods to `string`.
 * Rename `OrElse` overloads that return the item to `GetOrElse` which improves overload resolution.
 * Added `Curry`, `Uncurry` and `Flip` to the `Functional` Class
+* Remove `IToString`.