+++ changelog.md
@@ -34,4 +34,6 @@
   Old: `Result<int>.Ok(10)`. New: `Result.Ok(10)`.
 * Add `IndexOfOrNone`, `LastIndexOfOrNone`, `IndexOfAnyOrNone` and `LastIndexOfAnyOrNone` extension methods to `string`.
 * Rename `OrElse` overloads that return the item to `GetOrElse` which improves overload resolution.
-* Mark our functions as `[Pure]`
+* Added `Curry`, `Uncurry` and `Flip` to the `Functional` Class
+* Remove `IToString`.
+* Mark our functions as `[Pure]`.