+++ changelog.md
@@ -37,3 +37,5 @@
 * Added `Curry`, `Uncurry` and `Flip` to the `Functional` Class
 * Remove `IToString`.
 * Mark our functions as `[Pure]`.
+* Fix incorrect implementation on `Result.SelectMany` which called the `selectedResultSelector` even when the
+  result was an error. As a result (pun intended) of the fix, `ResultCombinationException` is no longer needed and also removed.