+++ changelog.md
@@ -29,7 +29,7 @@
 * Remove `Reader` monad based on `await`.
 * `Exception` created by `Result` monad contains valid stack trace
 
-## Unreleased
+## 2.0.0-rc.2
 * Move the `Ok` constructor of `Result<T>` to a non-generic class. This allows for the compiler to infer the generic type.
   Old: `Result<int>.Ok(10)`. New: `Result.Ok(10)`.
 * Add `IndexOfOrNone`, `LastIndexOfOrNone`, `IndexOfAnyOrNone` and `LastIndexOfAnyOrNone` extension methods to `string`.